# expression: implement vectorized evaluation for 'builtinLEIntSig'